# Optimize Pack for deep, ignored directories