--- a/internal/ignorefiles/terraformignore.go
+++ b/internal/ignorefiles/terraformignore.go
@@ -33,14 +33,14 @@ func readRules(input io.Reader) ([]rule, error) {
 		rule := rule{}
 		// Exclusions
 		if pattern[0] == '!' {
-			rule.excluded = true
+			rule.negated = true
 			pattern = pattern[1:]
-			// Mark all previous rules as having exclusions after it
+			// Mark all previous rules as having negations after it
 			for i := currentRuleIndex; i >= 0; i-- {
-				if rules[i].exclusionsAfter {
+				if rules[i].negationsAfter {
 					break
 				}
-				rules[i].exclusionsAfter = true
+				rules[i].negationsAfter = true
 			}
 		}
 		// If it is a directory, add ** so we catch descendants
@@ -67,11 +67,11 @@ func readRules(input io.Reader) ([]rule, error) {
 }
 
 type rule struct {
-	val             string         // the value of the rule itself
-	excluded        bool           // ! is present, an exclusion rule
-	exclusionsAfter bool           // exclusion rules appear after this rule
-	dirs            []string       // directories of the rule
-	regex           *regexp.Regexp // regular expression to match for the rule
+	val            string         // the value of the rule itself
+	negated        bool           // prefixed by !, a negated rule
+	negationsAfter bool           // negatied rules appear after this rule
+	dirs           []string       // directories of the rule
+	regex          *regexp.Regexp // regular expression to match for the rule
 }
 
 func (r *rule) match(path string) (bool, error) {
@@ -170,16 +170,16 @@ func (r *rule) compile() error {
 
 var defaultExclusions = []rule{
 	{
-		val:      strings.Join([]string{"**", ".git", "**"}, string(os.PathSeparator)),
-		excluded: false,
+		val:     strings.Join([]string{"**", ".git", "**"}, string(os.PathSeparator)),
+		negated: false,
 	},
 	{
-		val:      strings.Join([]string{"**", ".terraform", "**"}, string(os.PathSeparator)),
-		excluded: false,
+		val:     strings.Join([]string{"**", ".terraform", "**"}, string(os.PathSeparator)),
+		negated: false,
 	},
 	{
-		val:      strings.Join([]string{"**", ".terraform", "modules", "**"}, string(os.PathSeparator)),
-		excluded: true,
+		val:     strings.Join([]string{"**", ".terraform", "modules", "**"}, string(os.PathSeparator)),
+		negated: true,
 	},
 }