--- a/internal/ignorefiles/testdata/with-exclusion/.terraformignore
+++ b/internal/ignorefiles/testdata/with-exclusion/.terraformignore
@@ -1,0 +1,5 @@
+src/**/*
+# except at one directory
+!src/foo/bar.txt
+logs/
+tmp/