--- a/internal/ignorefiles/testdata/with-exclusion/src/baz/ignored.txt
+++ b/internal/ignorefiles/testdata/with-exclusion/src/baz/ignored.txt
@@ -1,0 +1,1 @@
+ignored