--- a/slug.go
+++ b/slug.go
@@ -199,15 +199,19 @@ func (p *Packer) packWalkFn(root, src, dst string, tarW *tar.Writer, meta *Meta,
 			return nil
 		}
 
-		if m := matchIgnoreRules(subpath, ignoreRules); m {
+		if r := matchIgnoreRules(subpath, ignoreRules); r.Match {
 			return nil
 		}
 
 		// Catch directories so we don't end up with empty directories,
 		// the files are ignored correctly
 		if info.IsDir() {
-			if m := matchIgnoreRules(subpath+string(os.PathSeparator), ignoreRules); m {
-				return nil
+			if r := matchIgnoreRules(subpath+string(os.PathSeparator), ignoreRules); r.Match {
+				if r.Dominating {
+					return filepath.SkipDir
+				} else {
+					return nil
+				}
 			}
 		}