--- a/testdata/archive-dir-no-external/.terraformignore
+++ b/testdata/archive-dir-no-external/.terraformignore
@@ -6,7 +6,7 @@
 
 # ignore a directory
 terraform.d/
-# exclude ignoring a directory at the root
+# negate ignoring a directory at the root
 !/terraform.d/
 # ignore a file at a subpath
 **/foo/bar.tf