… internal/ignorefiles/testdata/with-exclusion/logs/foo.txt
… internal/ignorefiles/testdata/with-exclusion/src/foo/bar.txt
… internal/ignorefiles/testdata/with-exclusion/tmp/tmp.txt