@brandonc 2023-11-23
@apparentlymart while I was pondering this change, I found a potentially problematic aspect of sourcebundle, where matching directories are deleted regardless of potential exclusion rules that follow it. I left a comment in the code.